feature gate some parts of common and make mods for them
