+++ common/Cargo.toml
@@ -8,12 +8,18 @@ celestia-types = {workspace = true}
 nmt-rs = {workspace = true}
 serde = {workspace = true}
 tendermint = {workspace = true}
-tendermint-proto = {workspace = true}
+tendermint-proto = {workspace = true, optional = true}
 sha3 = {workspace = true}
 thiserror = {workspace = true}
 alloy = {workspace = true, features = ["sol-types"]}
-prost = {workspace = true}
-tonic = {workspace = true}
+prost = {workspace = true, optional = true}
+tonic = {workspace = true, features = ["codegen", "prost"], optional = true}
 
 [build-dependencies]
-tonic-build = {workspace = true}
+tonic-build = {workspace = true, features=["prost"], optional = true}
+
+[features]
+default = ["grpc", "utils"]
+
+utils = ["tendermint-proto"]
+grpc = ["tonic", "prost", "tonic-build"]
