+++ common/build.rs
@@ -1,12 +1,17 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let code_gen_path = std::path::Path::new("src/generated");
-    if !code_gen_path.exists() {
-        std::fs::create_dir_all(code_gen_path)?;
+    #[cfg(feature = "grpc")]
+    {
+        let code_gen_path = std::path::Path::new("src/generated");
+        if !code_gen_path.exists() {
+            std::fs::create_dir_all(code_gen_path)?;
+        }
+        tonic_build::configure()
+            .build_server(true)
+            .build_client(true)
+            .out_dir(code_gen_path)
+            .compile_protos(&["proto/eqservice.proto"], &["proto/"])?;
+        Ok(())
     }
-    tonic_build::configure()
-        .build_server(true)
-        .build_client(true)
-        .out_dir(code_gen_path)
-        .compile_protos(&["proto/eqservice.proto"], &["proto/"])?;
+    #[cfg(not(feature = "grpc"))]
     Ok(())
 }
