+++ program-keccak-inclusion/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-eq-common = {path = "../common"}
+eq-common = {path = "../common", default-features = false}
 sp1-zkvm = {workspace = true}
 celestia-types = {workspace = true}
 nmt-rs = {workspace = true}
